% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.r
\name{device}
\alias{device}
\alias{device_get_count}
\alias{device_get_handle_by_index}
\alias{device_get_name}
\title{NVML Device Queries}
\usage{
device_get_count()

device_get_handle_by_index(index)

device_get_name(device)
}
\description{
Wrappers for queries NVML runs against a device.
}
\details{
The function naming scheme for the wrappers is:
\itemize{
  \item NVML name: \code{nvmlFunctionName()}
  \item nvsmi name: \code{function_name()}
}
}
\references{
\url{https://docs.nvidia.com/deploy/nvml-api/group__nvmlSystemQueries.html}
}
