% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_processes.r
\name{gpu_processes}
\alias{gpu_processes}
\title{gpu_processes}
\usage{
gpu_processes(type = "both")
}
\arguments{
\item{type}{One of \code{"compute"} (CUDA processes), \code{"graphics"} (OpenGL, etc), or
\code{"both"} (CUDA + OpenGL).}
}
\value{
A dataframe with one row per process found, and 5 columns:
\itemize{
  \item \code{GPU}: 
  \item \code{PID}: 
  \item \code{Type}: 
  \item \code{Process}: 
  \item \code{Memory}: 
}
}
\description{
Returns the
}
