% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{init}
\alias{init}
\alias{nvsmi_init}
\alias{nvsmi_shutdown}
\title{NVML Initialization and Cleanup}
\usage{
nvsmi_init()

nvsmi_shutdown()
}
\description{
Wrappers for NVML initialization and cleanup functions.
}
\details{
Unlike all other wrappers, the naming scheme is
\itemize{
  \item initialization:
  \itemize{
    \item NVML name: \code{nvmlInit()}
    \item nvsmi name: \code{nvsmi_init()}
  }
  \item shutdown:
  \itemize{
    \item NVML name: \code{nvmlShutdown()}
    \item nvsmi name: \code{nvsmi_shutdown()}
  }
}
}
\references{
\url{https://docs.nvidia.com/deploy/nvml-api/group__nvmlSystemQueries.html}
}
