% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{system}
\alias{system}
\alias{system_get_cuda_driver_version}
\alias{system_get_driver_version}
\alias{system_get_nvml_version}
\alias{system_get_process_name}
\title{NVML System Queries}
\usage{
system_get_cuda_driver_version()

system_get_driver_version()

system_get_nvml_version()

system_get_process_name(pid)
}
\arguments{
\item{pid}{Process id number.}
}
\description{
Wrappers for queries NVML runs against the local system.
}
\details{
The function naming scheme for the wrappers is:
\itemize{
  \item NVML name: \code{nvmlFunctionName()}
  \item nvsmi name: \code{function_name()}
}
}
\references{
\url{https://docs.nvidia.com/deploy/nvml-api/group__nvmlSystemQueries.html}
}
