% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi.r
\name{smi}
\alias{smi}
\title{smi}
\usage{
smi()
}
\value{
A list with 3 elements of class \code{nvidia_smi}:
\itemize{
  \item \code{version}: 
  \item \code{date}: 
  \item The \code{gpus} dataframe: 
  \itemize{
    \item \code{name}: 
    \item \code{busid}: 
    \item \code{persistence_mode}: 
    \item \code{disp}: 
    \item \code{speed}: 
    \item \code{temp}: 
    \item \code{perf}: 
    \item \code{power}: 
    \item \code{power_max}: 
    \item \code{memory_used}: 
    \item \code{memory_total}: 
    \item \code{utilization}: 
    \item \code{compute_mode}: 
  }
}
}
\description{
An \code{nvidia-smi} clone.
}
